backend/requirements.txt
# HexWard Backend Dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-socketio==5.11.0
python-multipart==0.0.9

# Database & Storage
sqlalchemy==2.0.31
alembic==1.13.2
sqlite3-to-sqlite==0.1.0
aiosqlite==0.20.0

# AI & Computer Vision
opencv-python==4.9.0.80
ultralytics==8.2.34
openai==1.30.1
pillow==10.3.0
numpy==1.26.4

# Audio Processing
whisper==1.1.10
sounddevice==0.4.6

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7